move: member 패키지 이동
member 패키지를 domain 패키지 아래로 이동
