+++ backend/src/main/java/com/metamong/mt/domain/member/exception/InvalidLoginRequestException.java
@@ -1,0 +1,10 @@
+package com.metamong.mt.domain.member.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class InvalidLoginRequestException extends RuntimeException {
+    private final InvalidLoginRequestType invalidLoginRequestType;
+}
