+++ backend/src/main/java/com/metamong/mt/domain/member/exception/InvalidLoginRequestType.java
@@ -1,0 +1,6 @@
+package com.metamong.mt.domain.member.exception;
+
+public enum InvalidLoginRequestType {
+    MEMBER_NOT_EXISTS,
+    PASSWORD_INCORRECT
+}
