+++ backend/src/main/java/com/metamong/mt/domain/member/exception/PasswordNotConfirmedException.java
@@ -1,0 +1,5 @@
+package com.metamong.mt.domain.member.exception;
+
+public class PasswordNotConfirmedException extends RuntimeException {
+
+}
