+++ backend/src/main/java/com/metamong/mt/domain/member/service/mailformatter/IdNotificationMailMessageFormatter.java
@@ -1,0 +1,22 @@
+package com.metamong.mt.domain.member.service.mailformatter;
+
+import org.springframework.stereotype.Component;
+
+import com.metamong.mt.global.mail.MailMessageFormatter;
+import com.metamong.mt.global.mail.MailType;
+
+@Component
+public class IdNotificationMailMessageFormatter extends MailMessageFormatter {
+
+    @Override
+    protected String getMessageFormat() {
+        return """
+                회원님의 아이디는 {0} 입니다.
+                """;
+    }
+
+    @Override
+    public MailType getSupportedMailType() {
+        return MailType.ID_NOTIFICATION;
+    }
+}
