--- backend/src/main/java/com/metamong/mt/member/dto/request/LoginRequestDto.java
@@ -1,16 +1,0 @@
-package com.metamong.mt.member.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequestDto {
-
-	@NotBlank(message = "아이디는 필수입니다.")
-    private String userId;
-
-	@NotBlank(message = "비밀번호는 필수입니다.")
-    private String password;
-}
