--- backend/src/main/java/com/metamong/mt/member/exception/InvalidLoginRequestException.java
@@ -1,10 +1,0 @@
-package com.metamong.mt.member.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class InvalidLoginRequestException extends RuntimeException {
-    private final InvalidLoginRequestType invalidLoginRequestType;
-}
