--- backend/src/main/java/com/metamong/mt/member/exception/InvalidLoginRequestType.java
@@ -1,6 +1,0 @@
-package com.metamong.mt.member.exception;
-
-public enum InvalidLoginRequestType {
-    MEMBER_NOT_EXISTS,
-    PASSWORD_INCORRECT
-}
