--- backend/src/main/java/com/metamong/mt/member/exception/PasswordNotConfirmedException.java
@@ -1,5 +1,0 @@
-package com.metamong.mt.member.exception;
-
-public class PasswordNotConfirmedException extends RuntimeException {
-
-}
