--- backend/src/main/java/com/metamong/mt/member/model/IdOrPw.java
@@ -1,6 +1,0 @@
-package com.metamong.mt.member.model;
-
-public enum IdOrPw {
-    ID,
-    PW
-}
